+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# 2.0.1
+  - Correctly merge multiple line with the multiline codec ref: #24
 # 2.0.0
   - Add support for stream identity, the ID will be generated from beat.id+resource_id or beat.name + beat.source if not present #22 #13
     The identity allow the multiline codec to correctly merge string from multiples files.